Add comments to QueryComponent. Extract inner test class for added test and use a singleton for test ShardRequest.
