+++ solr/core/src/test/org/apache/solr/handler/component/MockResponseBuilder.java
@@ -1,0 +1,54 @@
+package org.apache.solr.handler.component;
+
+import org.apache.solr.common.params.ShardParams;
+import org.apache.solr.common.params.SolrParams;
+import org.apache.solr.common.util.NamedList;
+import org.apache.solr.request.SolrQueryRequest;
+import org.apache.solr.response.SolrQueryResponse;
+import org.apache.solr.schema.IndexSchema;
+import org.apache.solr.schema.SchemaField;
+import org.apache.solr.schema.StrField;
+import org.apache.solr.search.SortSpec;
+import org.mockito.Mockito;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class MockResponseBuilder extends ResponseBuilder {
+
+    private MockResponseBuilder(SolrQueryRequest request, SolrQueryResponse response, List<SearchComponent> components) {
+        super(request, response, components);
+    }
+
+    public static MockResponseBuilder create() {
+
+        // the mocks
+        SolrQueryRequest request = Mockito.mock(SolrQueryRequest.class);
+        SolrQueryResponse response = Mockito.mock(SolrQueryResponse.class);
+        IndexSchema indexSchema = Mockito.mock(IndexSchema.class);
+        SolrParams params = Mockito.mock(SolrParams.class);
+
+        // SchemaField must be concrete due to field access
+        SchemaField uniqueIdField = new SchemaField("id", new StrField());
+
+        // we need this because QueryComponent adds a property to it.
+        NamedList<Object> responseHeader = new NamedList<>();
+
+        // the mock implementations
+        Mockito.when(request.getSchema()).thenReturn(indexSchema);
+        Mockito.when(indexSchema.getUniqueKeyField()).thenReturn(uniqueIdField);
+        Mockito.when(params.getBool(ShardParams.SHARDS_INFO)).thenReturn(false);
+        Mockito.when(request.getParams()).thenReturn(params);
+        Mockito.when(response.getResponseHeader()).thenReturn(responseHeader);
+
+        List<SearchComponent> components = new ArrayList<>();
+        return new MockResponseBuilder(request, response, components);
+
+    }
+
+    public MockResponseBuilder withSortSpec(SortSpec sortSpec) {
+        this.setSortSpec(sortSpec);
+        return this;
+    }
+
+}
